updates test and function
