+++ src/App.test.tsx
@@ -1,18 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import App from './App';
-import { getChange } from './helpers';
-
-// Denominations available: 1, 5, 10, 25, 50, 100
 
 test('renders learn react link', () => {
   render(<App />);
   const headEl = screen.getByText(/Hello World/i);
   expect(headEl).toBeTruthy();
-});
-
-test('getChange(1,1) should equal [] - an empty array', () => {
-  let totalCost = 5;
-  let amountPaid = 5;
-  let result = getChange(totalCost, amountPaid);
-  expect(result).toEqual([]);
 });
